# Faster wrapping multiplication