--- a/src/uint/mul/schoolbook.rs
+++ b/src/uint/mul/schoolbook.rs
@@ -40,6 +40,7 @@ pub const fn mul_wide(lhs: &[Limb], rhs: &[Limb], lo: &mut [Limb], hi: &mut [Lim
 }
 
 /// Add the schoolbook product of two limb slices to a limb slice, returning the carry.
+#[cfg(feature = "alloc")]
 #[inline]
 pub const fn carrying_add_mul(lhs: &[Limb], rhs: &[Limb], out: &mut [Limb]) -> Limb {
     assert!(