@andrewwhitehead 2025-09-04
@daxpedda The benchmarks are from a 2021 Macbook Pro M1. I've updated the numbers to include `wrapping_square`. It wouldn't hurt to have more benchmarks on other platforms but some of the tests didn't exist before.