@daxpedda 2025-09-04
That's very cool.
But I assume you meant to ping @dvdplm.